requirements.txt
absl-py==0.9.0
astor==0.8.1
astroid==2.3.3
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
chromedriver-binary==79.0.3945.36.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.2
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.2
mypy==0.740
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.2
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandocfilters==1.4.2
parso==0.5.2
pickleshare==0.7.5
Pillow==6.2.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
Pygments==2.5.2
pylint==2.4.2
pyparsing==2.4.6
pyrsistent==0.15.6
PySocks==1.7.1
python-dateutil==2.8.1
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.13.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
tweepy==3.8.0
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.7
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0